Improve loading of single-image files.
I was messing with this the other day (instead of working), trying to
load some of the art assets in Rebecca Heineman's 3DO port of Doom, and
found that they wouldn't load.  After looking at the files, it seems
that *some people* thought you could write chunks in any order.

Alter the loading logic such that, for files containing a single image,
defer decoding the pixel data until we hit EOF, when we hopefully have
all the chunks we need.  ANIM files continue to be processed as before.

Also:
- Add `static` qualifier to inline functions so the linker will stop
  trying to do something with them.
- New entries to .gitignore.
- Good $(GOD), has it really been 11 years since I wrote this?

=== History ===

3DO cel files are loosely inspired by Amiga IFF files.  One perennial
problem with writing Amiga IFF image loaders was that the various data
chunks could appear in any order, so loaders had to be quite flexible.
However, the FORM ILBM spec stated outright that the BODY (pixel data)
chunk had to come last, because you couldn't properly interpret it until
all the metadata (header, color map, color ranges, Amiga display mode
bits, etc.) was present.  I wrote this parser (apparently incorrectly)
assuming the same rule applied to 3DO cel files.  In the case of the cel
files in Doom, the PLUT appeared after the PDAT.